Change type for line number

+++ Team00/Code00/src/spa/src/Tokenizer/Token/PqlToken.h
@@ -18,16 +18,16 @@ enum class TokenType {
 
 class PqlToken {
 public:
-    PqlToken(TokenType, std::string, int);
+    PqlToken(TokenType, std::string, size_t);
 
     [[nodiscard]] TokenType GetTokenType() const;
     [[nodiscard]] std::string GetToken() const;
-    [[nodiscard]] int GetLineNumber() const;
+    [[nodiscard]] size_t GetLineNumber() const;
 
 private:
     TokenType type_;
     std::string token_;
-    int line_number_;
+    size_t line_number_;
 };
 
 #endif
